// apps/desktop/src/types/generated/CardDetail.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { Activity } from "./Activity";
import type { AgentTag } from "./AgentTag";
import type { CardRef } from "./CardRef";
import type { CardView } from "./CardView";
import type { Column } from "./Column";
import type { Comment } from "./Comment";

export type CardDetail = { card: CardView, column: Column, comments: Array<Comment>, activity: Array<Activity>, dependsOn: Array<CardRef>, dependents: Array<CardRef>, children: Array<CardRef>, agents: Array<AgentTag>, };